Move dictionary into primitives
