+++ library/Rattletrap/PropertyValue.hs
@@ -1,6 +1,6 @@
 module Rattletrap.PropertyValue where
 
-import Rattletrap.Dictionary
+import Rattletrap.Primitive.Dictionary
 import Rattletrap.Float32
 import Rattletrap.Int32
 import Rattletrap.List
